// pfe_frontend/src/app/models/employe.model.ts
export interface Employe {
  matricule: number;
  idWrd: string;
  name: string;
  departement: string;
  category: string;
  currentJob: string;
  startDate: string;
  projet: string;
  team: string;
  poste: string;
  sexe: string;
  level: string;
  maritalStatut: string;
  dateAncienne: string;
  dateNaissance: string;
  zrfz: string;
  zfgrq: number;
  trancheAge: string;
  telephone: string;
  destination: string;
}